[FIX] web: remove white spaces in ttf file names
Since 8ddb46a7af, some google fonts are included. The "Open Sans" fonts
prevents RPM packaging to work because of white spaces in file names. This
is a Python setuptools bug (https://bugs.python.org/issue809163) which is
not yet fixed.

With this commit, the "Open Sans" directory and ttf files are renamed
into "Open_Sans", solving the packaging problem.

closes odoo/odoo#35661

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>