[REF] web: make ribbon widget great and user friendly
PURPOSE

Have a nice ribbon widget to display state of records: part II

SPECIFICATIONS

A POC for a new ribbon widget has been recently introduced at 4cafdbc2a635df99ae8e4f1fd2904582ed732b10.
It adds a ribbon on the top right of a form to make some status more visible
and more user friendly. This commit greatly improves the design of this
ribbon in order to have a real ribbon.

You have to specify the text attribute to show the text on the ribbon.
You can also specify a bg_color attribute to color the background with a
bootstrap background color class according to the business meaning of the
flag displayed in the ribbon.

LINKS

Task ID : 2032621
PR : #34662