[FIX] sale_stock: wrong route computed on so lines
_compute_is_mto was based on the sequence of the routes,
as most of those sequences were hardcoded, some had the same values,
resulting in wrong calculations. Furthermore, the method was setting the so lines
as mto if the sequence of the MTO route had the priority on the other ones, even if
the mto route wasn't set on the product.

This simply ensure to check the mto route is set on the product,
sequences should be set manually to prevent any sequence conflict.

TaskID: 2053703

closes odoo/odoo#35740

Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>