[FIX] account, payment, website_sale: QR code based on partner's country
The QR code was limited to a set of countries listed from Wikipedia [1],
but the list is incomplete. Any partner from a SEPA country should have
the QR code printed since we don't know which banks support it.

At the moment we simply remove the limitation, and hopefully a proper
filtering will be done in master.

[1] https://en.wikipedia.org/wiki/EPC_QR_code

opw-2052315

closes odoo/odoo#35827

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>