[FIX] delivery: wrong domain on carrier list
When you try to find a carrier in the wizard that let you add one on
your sale order, you don't see all of them because of a domain that only
show the ones that havethe same company id of the sale order, but it
should also show you the ones that don't have company_id set.

So we removed the domain in the view, and the invisible company_id as it
is not used anymore. Because the domain we need is already applied in
'available_carrier_ids'.

This behavior has been introduced in rev: 8baefb08a048ef69c761c679eaee1e4fe3ba4cee

closes odoo/odoo#35856

Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>