[IMP] sale: reflect down payment changes in the SO
Allow users to modify a down payment invoice's properly by reflecting
the change done to the invoice (e.g. taxes, amount) on the initial
downpayment SO line when the invoice is posted. This support several
nice flows, such a allowing (partial and full) refunds of downpayments
while preserving still a meaningful 'last invoice with deducted down
payments' amount that sums up to the total of the order.

Note that this does not apply to locked orders (by definition of the
term 'locked').

closes odoo/odoo#35982

Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>